+++ examples/render-opengl/src/main.rs
@@ -83,23 +83,6 @@ fn main() -> Result<(), Box<dyn Error>> {
                 puppet.set_named_param("Head:: Yaw-Pitch", Vec2::new(t.cos(), 0.0));
                 puppet.end_set_params();
 
-                // TODO: this for loop won't compile (even if we clone all the driver UUIDs)
-                // because Rust is not bad enough to allow an OOP architecture :(
-                //
-                for driver_uuid in puppet.drivers.clone() {
-                    let Some(driver) = puppet.nodes.get_node_mut(driver_uuid) else {
-                        continue;
-                    };
-
-                    match &mut driver.data {
-                        InoxData::SimplePhysics(system) => {
-                            let nrc = &puppet.render_ctx.node_render_ctxs[&driver.uuid];
-                            system.update_driver(0.01, nrc, &mut puppet, system.param);
-                        }
-                        _ => (),
-                    }
-                }
-
                 renderer.render(&puppet);
 
                 gl_surface.swap_buffers(&gl_ctx).unwrap();
