+++ inox2d/src/nodes/mod.rs
@@ -1,4 +1,3 @@
 pub mod node;
 pub mod node_data;
 pub mod node_tree;
-pub mod physics;
