+++ inox2d/src/physics/pendulum.rs
@@ -1,40 +1,9 @@
+use super::{
+    runge_kutta::{self, PhysicsState, PhysicsSystem},
+    SimplePhysics, SimplePhysicsProps,
+};
+
 use glam::Vec2;
-
-use crate::nodes::physics::{self, SimplePhysics, SimplePhysicsProps};
-
-#[derive(Clone, Debug)]
-pub struct PhysicsState<const N: usize> {
-    pub vars: [f32; N],
-    pub derivatives: [f32; N],
-    pub t: f32,
-}
-
-impl<const N: usize> Default for PhysicsState<N> {
-    fn default() -> Self {
-        Self {
-            vars: [0.; N],
-            derivatives: [0.; N],
-            t: 0.,
-        }
-    }
-}
-
-pub trait PhysicsSystem<const N: usize> {
-    type Output;
-
-    fn state(&self) -> &PhysicsState<N>;
-    fn state_mut(&mut self) -> &mut PhysicsState<N>;
-    fn set_state(&mut self, state: PhysicsState<N>);
-
-    fn eval(&mut self, physics_props: &SimplePhysicsProps, t: f32) -> &mut PhysicsState<N>;
-    fn tick(&mut self, node: &SimplePhysics, h: f32) -> Self::Output;
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ParamMapMode {
-    AngleLength,
-    XY,
-}
 
 #[derive(Default, Debug, Clone)]
 pub struct Pendulum {
@@ -71,8 +40,6 @@ impl Pendulum {
 }
 
 impl PhysicsSystem<2> for Pendulum {
-    type Output = Vec2;
-
     fn state(&self) -> &PhysicsState<2> {
         &self.physics_state
     }
@@ -98,31 +65,22 @@ impl PhysicsSystem<2> for Pendulum {
 
         &mut self.physics_state
     }
+}
 
-    // TODO: oh my god there are 2 tick functions. This really needs a rename/refactor/whatever.
-
-    fn tick(&mut self, node: &SimplePhysics, h: f32) -> Vec2 {
+impl Pendulum {
+    pub fn tick(&mut self, anchor: &Vec2, props: &SimplePhysicsProps, dt: f32) -> Vec2 {
         // Compute the angle against the updated anchor position
-        let delta_bob = self.bob - node.anchor;
+        let delta_bob = self.bob - *anchor;
         self.set_angle(f32::atan2(-delta_bob.x, delta_bob.y));
 
         // Run the pendulum simulation in terms of angle
-        physics::tick(self, &node.props, h);
+        runge_kutta::tick(self, props, dt);
 
         // Update bob's position at the new angle
         let angle = self.angle();
         let delta_bob = Vec2::new(-angle.sin(), angle.cos());
-        self.bob = node.anchor + delta_bob * node.props.length;
+        self.bob = *anchor + delta_bob * props.length;
 
         self.bob
     }
-}
-
-/// Physics model to use for simple physics
-#[derive(Debug, Clone)]
-pub enum SimplePhysicsSystem {
-    /// Rigid pendulum
-    Pendulum(Pendulum),
-    // TODO: Springy pendulum
-    // SpringPendulum(),
 }
