+++ inox2d/src/physics/simple_physics.rs
@@ -1,54 +1,13 @@
 use std::f32::consts::PI;
-use std::ops::Not;
 
 use glam::{vec2, vec4, Vec2};
 
+use super::{ParamMapMode, SimplePhysics, SimplePhysicsSystem};
 use crate::params::ParamUuid;
 use crate::puppet::{Puppet, PuppetPhysics};
 use crate::render::NodeRenderCtx;
-use crate::system::{ParamMapMode, PhysicsSystem, SimplePhysicsSystem};
-
-#[derive(Debug, Clone, Default)]
-pub struct SimplePhysicsProps {
-    /// Gravity scale (1.0 = puppet gravity)
-    pub gravity: f32,
-    /// Pendulum/spring rest length (pixels)
-    pub length: f32,
-    /// Resonant frequency (Hz)
-    pub frequency: f32,
-    /// Angular damping ratio
-    pub angle_damping: f32,
-    /// Length damping ratio
-    pub length_damping: f32,
-
-    pub output_scale: Vec2,
-}
-
-#[derive(Debug, Clone)]
-pub struct SimplePhysics {
-    pub param: ParamUuid,
-
-    pub system: SimplePhysicsSystem,
-    pub map_mode: ParamMapMode,
-
-    pub offset_props: SimplePhysicsProps,
-    pub props: SimplePhysicsProps,
-
-    /// Whether physics system listens to local transform only.
-    pub local_only: bool,
-
-    pub anchor: Vec2,
-    pub output: Vec2,
-}
 
 impl SimplePhysics {
-    pub fn tick_system(&mut self, h: f32) {
-        match &mut self.system {
-            SimplePhysicsSystem::Pendulum(system) => tick(system, &self.props, h),
-            // spring pendulum when
-        }
-    }
-
     fn update_inputs(&mut self, node_render_ctx: &NodeRenderCtx) {
         let anchor_pos = match self.local_only {
             true => node_render_ctx.trans_offset.translation.extend(1.0),
@@ -58,12 +17,7 @@ impl SimplePhysics {
         self.anchor = vec2(anchor_pos.x, anchor_pos.y);
     }
 
-    fn update_outputs(
-        &mut self,
-        node_render_ctx: &NodeRenderCtx,
-        puppet: &mut Puppet,
-        param_uuid: ParamUuid,
-    ) {
+    fn calc_outputs(&mut self, node_render_ctx: &NodeRenderCtx) -> Vec2 {
         let oscale = self.final_output_scale();
 
         // "Okay, so this is confusing. We want to translate the angle back to local space, but not the coordinates."
@@ -94,16 +48,15 @@ impl SimplePhysics {
             }
         };
 
-        puppet.set_param(param_uuid, param_value * oscale);
+        param_value * oscale
     }
 
-    pub fn update_driver(
+    pub fn update(
         &mut self,
         dt: f32,
         node_render_ctx: &NodeRenderCtx,
-        puppet: &mut Puppet,
         param_uuid: ParamUuid,
-    ) {
+    ) -> Vec2 {
         // Timestep is limited to 10 seconds.
         // If you're getting 0.1 FPS, you have bigger issues to deal with.
         let mut h = dt.min(10.);
@@ -112,13 +65,13 @@ impl SimplePhysics {
 
         // Minimum physics timestep: 0.01s
         while h > 0.01 {
-            self.tick_system(0.01);
+            self.system.tick(&self.anchor, &self.props, 0.01);
             h -= 0.01;
         }
 
-        self.tick_system(h);
+        self.system.tick(&self.anchor, &self.props, h);
 
-        self.update_outputs(node_render_ctx, puppet, param_uuid);
+        self.calc_outputs(node_render_ctx)
     }
 
     pub fn update_anchor(&mut self) {
@@ -155,50 +108,4 @@ impl SimplePhysics {
     pub fn final_output_scale(&self) -> Vec2 {
         self.props.output_scale * self.offset_props.output_scale
     }
-}
-
-pub fn tick<const N: usize, P: PhysicsSystem<N>>(
-    system: &mut P,
-    physics_props: &SimplePhysicsProps,
-    h: f32,
-) {
-    let curs;
-    let t = {
-        let phys = system.state_mut();
-        curs = phys.vars;
-        phys.derivatives = [0.; N];
-        phys.t
-    };
-
-    let phys = system.eval(physics_props, t);
-    let k1s = phys.derivatives;
-
-    for i in 0..N {
-        phys.vars[i] = curs[i] + h * k1s[i] / 2.;
-    }
-    let phys = system.eval(physics_props, t + h / 2.);
-    let k2s = phys.derivatives;
-
-    for i in 0..N {
-        phys.vars[i] = curs[i] + h * k2s[i] / 2.;
-    }
-    let phys = system.eval(physics_props, t + h / 2.);
-    let k3s = phys.derivatives;
-
-    for i in 0..N {
-        phys.vars[i] = curs[i] + h * k3s[i];
-    }
-    let phys = system.eval(physics_props, t + h);
-    let k4s = phys.derivatives;
-
-    for i in 0..N {
-        phys.vars[i] = curs[i] + h * (k1s[i] + 2. * k2s[i] + 2. * k3s[i] + k4s[i]) / 6.;
-        if phys.vars[i].is_finite().not() {
-            // Simulation failed, revert
-            phys.vars = curs;
-            break;
-        }
-    }
-
-    phys.t += h;
 }
